Authentication/requirements.txt

# User authentication packages
alembic
aniso8601
appdirs
asgiref
bcrypt
black
celery
cryptography
fastapi
graphene==2.1.9
graphene-sqlalchemy==2.1.0
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet
h11
Mako
MarkupSafe
mypy-extensions
pathspec
promise
psycopg2-binary
pycparser
pydantic
PyJWT
python-dateutil
python-dotenv
python-editor
regex
Rx
singledispatch
six
SQLAlchemy
starlette
starlette-graphql
toml
typing-extensions
uvicorn>=0.15.0
